ai-service/requirements.txt
# InterviewOS AI Service — Python dependencies

assemblyai>=0.30.0
openai>=1.30.0          # used to call NVIDIA NIM (OpenAI-compatible)
fastapi>=0.111.0
uvicorn>=0.30.0
python-multipart>=0.0.9
pydantic>=2.7.0
python-dotenv>=1.0.0
httpx>=0.27.0
